Moved ApplicationDom into its own subdir.
