+++ lib/web_ui/lib/src/engine.dart
@@ -170,11 +170,11 @@ export 'engine/text_editing/text_editing.dart';
 export 'engine/util.dart';
 export 'engine/validators.dart';
 export 'engine/vector_math.dart';
-export 'engine/view_embedder/application_dom.dart';
-export 'engine/view_embedder/custom_element_application_dom.dart';
+export 'engine/view_embedder/application_dom/application_dom.dart';
+export 'engine/view_embedder/application_dom/custom_element_application_dom.dart';
+export 'engine/view_embedder/application_dom/full_page_application_dom.dart';
 export 'engine/view_embedder/dimensions_provider/custom_element_dimensions_provider.dart';
 export 'engine/view_embedder/dimensions_provider/dimensions_provider.dart';
 export 'engine/view_embedder/dimensions_provider/full_page_dimensions_provider.dart';
-export 'engine/view_embedder/full_page_application_dom.dart';
 export 'engine/view_embedder/hot_restart_cache_handler.dart';
 export 'engine/window.dart';
